Estimate the false discovery proportion of a record-linkage method via synthetic augmentation

Repeatedly augments file B with synthetic records (synthesise()), runs the chosen record-linkage method link_with_*, and compares the synthetic-vs-real proportion among linked pairs to estimate the false discovery proportion, for a range of score thresholds (0.50 to 0.99). See https://doi.org/10.1002/sim.70292 for the method.

compute_augmRL_FDP_synth(
  synth_method,
  encodedA,
  encodedB,
  PIVs,
  n_synth = NULL,
  restrict_support_intersection = TRUE,
  maxIter4CV = 10,
  n_repeats = 10,
  RL_method,
  ...
)

Arguments

synth_method

Passed to synthesise(): "arf", "synthpop", or "mice".

encodedA, encodedB

The two encoded data sources as returned by prepare_data() (encodedA is the smaller one).

PIVs

Character vector, names of the PIVs.

n_synth

Integer, number of synthetic records to generate per iteration (default: 10% of nrow(encodedB)).

restrict_support_intersection

Passed to synthesise().

maxIter4CV

Integer, max number of retries per iteration if no valid FDP estimate is obtained.

n_repeats

Integer, number of augmentation iterations to average over.

RL_method

One of "multilink", "fastLink", "BRL", "reclin2", "diyar", "fedmatch", "FlexRL".

...

Extra arguments forwarded to the chosen link_with_* wrapper (i.e. to the underlying record-linkage package).

Value

List with FDP_score_estimator, FDP_synth_estimator, Linked_pairs_augm, Linked_pairs: data frames (n_repeats rows x 50 thresholds, 0.50 to 0.99).

Examples

PIVs_config <- list( V1 = list(dynamics = "stable",
                               bound_mistakes = c(0.10,0.10),
                               fix_mistakes = c(NA,NA)),
                     V2 = list(dynamics = "stable",
                               bound_mistakes = c(0.10,0.10),
                               fix_mistakes = c(NA,NA)),
                     V3 = list(dynamics = "flexible",
                               bound_mistakes = c(NA,NA),
                               fix_mistakes = c(NA,NA)),
                     V4 = list(dynamics = "structured",
                               bound_mistakes = c(NA,NA),
                               fix_mistakes = c(0.03,0.03),
                               cond_hazard_cov = list(cov1=c("Xe", "Xf"),
                                                      cov2=c())) )
n_values  <- c( 5, 6, 7, 12 )
p_mistake <- list( V1 = c(0.02, 0.02), V2 = c(0.02, 0.02),
                   V3 = c(0.05, 0.05), V4 = c(0.02, 0.02) )
p_missing <- list( V1 = c(0.005, 0.005), V2 = c(0.005, 0.005),
                   V3 = c(0.005, 0.005), V4 = c(0.005, 0.005) )
cond_hazard_params <- list( V1 = c(), V2 = c(), 
                             V3 = c(), V4 = log(c(0.7, 0.6, 0.5)) )
gen_data <- simulate_data( PIVs_config, n_values, c(250, 300), 200, 
                           p_mistake, p_missing, cond_hazard_params, TRUE )
prep_data <- prepare_data( gen_data$data1, gen_data$data2, "1", "2",
                           PIVs_config, TRUE, "entity_id", TRUE )
#> '2' is the larger source, saved as B; '1' saved as A.
PIVs <- names(PIVs_config)                         
compute_augmRL_FDP_synth( "arf", prep_data$encodedA, prep_data$encodedB, PIVs, 
                          NULL, TRUE, 1, 2, "BRL", flds = PIVs, 
                          types = rep("bi",length(PIVs)) )
#> Iteration: 0, Accuracy: 48.74%
#> Warning: PIV 'V1': synthetic records out of the common support were removed.
#> Warning: PIV 'V4': synthetic records out of the common support were removed.
#> Iteration: 0, Accuracy: 45.1%
#> Warning: PIV 'V1': synthetic records out of the common support were removed.
#> Warning: PIV 'V4': synthetic records out of the common support were removed.
#> BRL results (average over 2 iterations) (augmented RL task) `compute_augmRL_FDP_synth`
#>                            0.50  0.55  0.60  0.65  0.70  0.75  0.80  0.85 0.90
#> FDP model score estimator  0.20  0.19  0.19  0.19  0.19  0.18  0.16  0.13    0
#> FDP synth data estimator   0.05  0.06  0.06  0.06  0.06  0.07  0.09  0.00    0
#> Linked pairs (augm. RL)   92.00 90.00 90.00 90.00 86.00 78.00 52.00 20.00    0
#> Linked pairs (RL)         93.00 90.00 90.00 90.00 86.00 78.00 52.00 20.00    0
#>                           0.95
#> FDP model score estimator    0
#> FDP synth data estimator     0
#> Linked pairs (augm. RL)      0
#> Linked pairs (RL)            0
#> $FDP_score_estimator
#>        0.50      0.51      0.52      0.53      0.54      0.55      0.56
#> 1 0.2059693 0.1963248 0.1931481 0.1931481 0.1931481 0.1931481 0.1931481
#> 2 0.1964855 0.1932601 0.1932601 0.1901358 0.1901358 0.1901358 0.1901358
#>        0.57      0.58      0.59      0.60      0.61      0.62      0.63
#> 1 0.1931481 0.1931481 0.1931481 0.1931481 0.1931481 0.1931481 0.1931481
#> 2 0.1901358 0.1901358 0.1901358 0.1901358 0.1901358 0.1901358 0.1901358
#>        0.64      0.65      0.66      0.67      0.68      0.69      0.70
#> 1 0.1931481 0.1931481 0.1914856 0.1914856 0.1914856 0.1901263 0.1901263
#> 2 0.1901358 0.1901358 0.1901358 0.1901358 0.1886642 0.1857982 0.1830196
#>        0.71      0.72      0.73      0.74      0.75      0.76      0.77
#> 1 0.1889655 0.1878811 0.1846720 0.1827846 0.1819444 0.1795238 0.1723366
#> 2 0.1803614 0.1790921 0.1765972 0.1743305 0.1722807 0.1722807 0.1698174
#>        0.78      0.79     0.80      0.81      0.82      0.83      0.84
#> 1 0.1690278 0.1659815 0.159390 0.1547654 0.1511111 0.1454167 0.1399111
#> 2 0.1650912 0.1573180 0.153786 0.1487500 0.1456061 0.1400601 0.1361806
#>        0.85      0.86      0.87      0.88      0.89 0.90 0.91 0.92 0.93 0.94
#> 1 0.1349708 0.1300000 0.1253704 0.1177778 0.0000000    0    0    0    0    0
#> 2 0.1276768 0.1257778 0.1193939 0.1140000 0.1066667    0    0    0    0    0
#>   0.95 0.96 0.97 0.98 0.99
#> 1    0    0    0    0    0
#> 2    0    0    0    0    0
#> 
#> $FDP_synth_estimator
#>        0.50      0.51      0.52      0.53      0.54      0.55      0.56
#> 1 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> 2 0.1098901 0.1111111 0.1111111 0.1123596 0.1123596 0.1123596 0.1123596
#>        0.57      0.58      0.59      0.60      0.61      0.62      0.63
#> 1 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> 2 0.1123596 0.1123596 0.1123596 0.1123596 0.1123596 0.1123596 0.1123596
#>        0.64      0.65      0.66      0.67      0.68      0.69      0.70
#> 1 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> 2 0.1123596 0.1123596 0.1123596 0.1123596 0.1136364 0.1162791 0.1190476
#>        0.71      0.72      0.73      0.74      0.75      0.76      0.77
#> 1 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> 2 0.1219512 0.1234568 0.1265823 0.1298701 0.1333333 0.1333333 0.1388889
#>        0.78      0.79      0.80     0.81 0.82 0.83 0.84 0.85 0.86 0.87 0.88
#> 1 0.0000000 0.0000000 0.0000000 0.000000    0    0    0    0    0    0    0
#> 2 0.1515152 0.1754386 0.1886792 0.212766    0    0    0    0    0    0    0
#>   0.89 0.90 0.91 0.92 0.93 0.94 0.95 0.96 0.97 0.98 0.99
#> 1    0    0    0    0    0    0    0    0    0    0    0
#> 2    0    0    0    0    0    0    0    0    0    0    0
#> 
#> $Linked_pairs_augm
#>   0.50 0.51 0.52 0.53 0.54 0.55 0.56 0.57 0.58 0.59 0.60 0.61 0.62 0.63 0.64
#> 1   94   91   90   90   90   90   90   90   90   90   90   90   90   90   90
#> 2   91   90   90   89   89   89   89   89   89   89   89   89   89   89   89
#>   0.65 0.66 0.67 0.68 0.69 0.70 0.71 0.72 0.73 0.74 0.75 0.76 0.77 0.78 0.79
#> 1   90   89   89   89   88   88   87   86   83   81   80   77   68   64   60
#> 2   89   89   89   88   86   84   82   81   79   77   75   75   72   66   57
#>   0.80 0.81 0.82 0.83 0.84 0.85 0.86 0.87 0.88 0.89 0.90 0.91 0.92 0.93 0.94
#> 1   51   45   40   32   25   19   13    6    1    0    0    0    0    0    0
#> 2   53   47   44   37   32   22   20   11    5    1    0    0    0    0    0
#>   0.95 0.96 0.97 0.98 0.99
#> 1    0    0    0    0    0
#> 2    0    0    0    0    0
#> 
#> $Linked_pairs
#>   0.50 0.51 0.52 0.53 0.54 0.55 0.56 0.57 0.58 0.59 0.60 0.61 0.62 0.63 0.64
#> 1   94   91   90   90   90   90   90   90   90   90   90   90   90   90   90
#> 2   92   91   91   90   90   90   90   90   90   90   90   90   90   90   90
#>   0.65 0.66 0.67 0.68 0.69 0.70 0.71 0.72 0.73 0.74 0.75 0.76 0.77 0.78 0.79
#> 1   90   89   89   89   88   88   87   86   83   81   80   77   68   64   60
#> 2   90   90   90   89   87   85   83   82   80   78   76   76   73   67   58
#>   0.80 0.81 0.82 0.83 0.84 0.85 0.86 0.87 0.88 0.89 0.90 0.91 0.92 0.93 0.94
#> 1   51   45   40   32   25   19   13    6    1    0    0    0    0    0    0
#> 2   54   48   44   37   32   22   20   11    5    1    0    0    0    0    0
#>   0.95 0.96 0.97 0.98 0.99
#> 1    0    0    0    0    0
#> 2    0    0    0    0    0
#> 
compute_augmRL_FDP_synth( "arf", prep_data$encodedA, prep_data$encodedB, PIVs, 
                          NULL, TRUE, 1, 2, "FlexRL", data = prep_data, 
                          StEM_iter = 5, StEM_burnin = 2, 
                          gibbs_iter = 5, gibbs_burnin = 2, n_post_sample = 5 )
#> Iteration: 0, Accuracy: 45.1%
#> Warning: PIV 'V1': synthetic records out of the common support were removed.
#> Warning: PIV 'V4': synthetic records out of the common support were removed.
#> FlexRL
#> Iteration: 0, Accuracy: 47.47%
#> Warning: PIV 'V3': synthetic records out of the common support were removed.
#> Warning: PIV 'V4': synthetic records out of the common support were removed.
#> FlexRL
#> FlexRL results (average over 2 iterations) (augmented RL task) `compute_augmRL_FDP_synth`
#>                             0.50   0.55   0.60   0.65   0.70   0.75  0.80  0.85
#> FDP model score estimator   0.13   0.13   0.05   0.05   0.05   0.05  0.00  0.00
#> FDP synth data estimator    0.63   0.63   0.40   0.40   0.40   0.40  0.34  0.34
#> Linked pairs (augm. RL)   143.00 143.00 112.00 112.00 112.00 112.00 86.00 86.00
#> Linked pairs (RL)         152.00 152.00 116.00 116.00 116.00 116.00 89.00 89.00
#>                            0.90  0.95
#> FDP model score estimator  0.00  0.00
#> FDP synth data estimator   0.34  0.34
#> Linked pairs (augm. RL)   86.00 86.00
#> Linked pairs (RL)         89.00 89.00
#> $FDP_score_estimator
#>        0.50      0.51      0.52      0.53      0.54      0.55      0.56
#> 1 0.1241830 0.1241830 0.1241830 0.1241830 0.1241830 0.1241830 0.1241830
#> 2 0.1350993 0.1350993 0.1350993 0.1350993 0.1350993 0.1350993 0.1350993
#>        0.57      0.58      0.59       0.60       0.61       0.62       0.63
#> 1 0.1241830 0.1241830 0.1241830 0.03931624 0.03931624 0.03931624 0.03931624
#> 2 0.1350993 0.1350993 0.1350993 0.05517241 0.05517241 0.05517241 0.05517241
#>         0.64       0.65       0.66       0.67       0.68       0.69       0.70
#> 1 0.03931624 0.03931624 0.03931624 0.03931624 0.03931624 0.03931624 0.03931624
#> 2 0.05517241 0.05517241 0.05517241 0.05517241 0.05517241 0.05517241 0.05517241
#>         0.71       0.72       0.73       0.74       0.75       0.76       0.77
#> 1 0.03931624 0.03931624 0.03931624 0.03931624 0.03931624 0.03931624 0.03931624
#> 2 0.05517241 0.05517241 0.05517241 0.05517241 0.05517241 0.05517241 0.05517241
#>         0.78       0.79 0.80 0.81 0.82 0.83 0.84 0.85 0.86 0.87 0.88 0.89 0.90
#> 1 0.03931624 0.03931624    0    0    0    0    0    0    0    0    0    0    0
#> 2 0.05517241 0.05517241    0    0    0    0    0    0    0    0    0    0    0
#>   0.91 0.92 0.93 0.94 0.95 0.96 0.97 0.98 0.99
#> 1    0    0    0    0    0    0    0    0    0
#> 2    0    0    0    0    0    0    0    0    0
#> 
#> $FDP_synth_estimator
#>        0.50      0.51      0.52      0.53      0.54      0.55      0.56
#> 1 0.5517241 0.5517241 0.5517241 0.5517241 0.5517241 0.5517241 0.5517241
#> 2 0.7092199 0.7092199 0.7092199 0.7092199 0.7092199 0.7092199 0.7092199
#>        0.57      0.58      0.59      0.60      0.61      0.62      0.63
#> 1 0.5517241 0.5517241 0.5517241 0.4464286 0.4464286 0.4464286 0.4464286
#> 2 0.7092199 0.7092199 0.7092199 0.3571429 0.3571429 0.3571429 0.3571429
#>        0.64      0.65      0.66      0.67      0.68      0.69      0.70
#> 1 0.4464286 0.4464286 0.4464286 0.4464286 0.4464286 0.4464286 0.4464286
#> 2 0.3571429 0.3571429 0.3571429 0.3571429 0.3571429 0.3571429 0.3571429
#>        0.71      0.72      0.73      0.74      0.75      0.76      0.77
#> 1 0.4464286 0.4464286 0.4464286 0.4464286 0.4464286 0.4464286 0.4464286
#> 2 0.3571429 0.3571429 0.3571429 0.3571429 0.3571429 0.3571429 0.3571429
#>        0.78      0.79      0.80      0.81      0.82      0.83      0.84
#> 1 0.4464286 0.4464286 0.5617978 0.5617978 0.5617978 0.5617978 0.5617978
#> 2 0.3571429 0.3571429 0.1204819 0.1204819 0.1204819 0.1204819 0.1204819
#>        0.85      0.86      0.87      0.88      0.89      0.90      0.91
#> 1 0.5617978 0.5617978 0.5617978 0.5617978 0.5617978 0.5617978 0.5617978
#> 2 0.1204819 0.1204819 0.1204819 0.1204819 0.1204819 0.1204819 0.1204819
#>        0.92      0.93      0.94      0.95      0.96      0.97      0.98
#> 1 0.5617978 0.5617978 0.5617978 0.5617978 0.5617978 0.5617978 0.5617978
#> 2 0.1204819 0.1204819 0.1204819 0.1204819 0.1204819 0.1204819 0.1204819
#>        0.99
#> 1 0.5617978
#> 2 0.1204819
#> 
#> $Linked_pairs_augm
#>   0.50 0.51 0.52 0.53 0.54 0.55 0.56 0.57 0.58 0.59 0.60 0.61 0.62 0.63 0.64
#> 1  145  145  145  145  145  145  145  145  145  145  112  112  112  112  112
#> 2  141  141  141  141  141  141  141  141  141  141  112  112  112  112  112
#>   0.65 0.66 0.67 0.68 0.69 0.70 0.71 0.72 0.73 0.74 0.75 0.76 0.77 0.78 0.79
#> 1  112  112  112  112  112  112  112  112  112  112  112  112  112  112  112
#> 2  112  112  112  112  112  112  112  112  112  112  112  112  112  112  112
#>   0.80 0.81 0.82 0.83 0.84 0.85 0.86 0.87 0.88 0.89 0.90 0.91 0.92 0.93 0.94
#> 1   89   89   89   89   89   89   89   89   89   89   89   89   89   89   89
#> 2   83   83   83   83   83   83   83   83   83   83   83   83   83   83   83
#>   0.95 0.96 0.97 0.98 0.99
#> 1   89   89   89   89   89
#> 2   83   83   83   83   83
#> 
#> $Linked_pairs
#>   0.50 0.51 0.52 0.53 0.54 0.55 0.56 0.57 0.58 0.59 0.60 0.61 0.62 0.63 0.64
#> 1  153  153  153  153  153  153  153  153  153  153  117  117  117  117  117
#> 2  151  151  151  151  151  151  151  151  151  151  116  116  116  116  116
#>   0.65 0.66 0.67 0.68 0.69 0.70 0.71 0.72 0.73 0.74 0.75 0.76 0.77 0.78 0.79
#> 1  117  117  117  117  117  117  117  117  117  117  117  117  117  117  117
#> 2  116  116  116  116  116  116  116  116  116  116  116  116  116  116  116
#>   0.80 0.81 0.82 0.83 0.84 0.85 0.86 0.87 0.88 0.89 0.90 0.91 0.92 0.93 0.94
#> 1   94   94   94   94   94   94   94   94   94   94   94   94   94   94   94
#> 2   84   84   84   84   84   84   84   84   84   84   84   84   84   84   84
#>   0.95 0.96 0.97 0.98 0.99
#> 1   94   94   94   94   94
#> 2   84   84   84   84   84
#>