Estimate the false discovery proportion of a record-linkage method via synthetic augmentation

Repeatedly augments file B with synthetic records (synthesise()), runs the chosen record-linkage method link_with_*, and compares the synthetic-vs-real proportion among linked pairs to estimate the false discovery proportion, for a range of score thresholds (0.50 to 0.99). See https://doi.org/10.1002/sim.70292 for the method.

compute_augmRL_FDP_synth(
  synth_method,
  encodedA,
  encodedB,
  PIVs,
  n_synth = NULL,
  restrict_support_intersection = TRUE,
  maxIter4CV = 10,
  n_repeats = 10,
  RL_method,
  ...
)

Arguments

synth_method

Passed to synthesise(): "arf", "synthpop", or "mice".

encodedA, encodedB

The two encoded data sources as returned by prepare_data() (encodedA is the smaller one).

PIVs

Character vector, names of the PIVs.

n_synth

Integer, number of synthetic records to generate per iteration (default: 10% of nrow(encodedB)).

restrict_support_intersection

Passed to synthesise().

maxIter4CV

Integer, max number of retries per iteration if no valid FDP estimate is obtained.

n_repeats

Integer, number of augmentation iterations to average over.

RL_method

One of "multilink", "fastLink", "BRL", "reclin2", "diyar", "fedmatch", "FlexRL".

...

Extra arguments forwarded to the chosen link_with_* wrapper (i.e. to the underlying record-linkage package).

Value

List with FDP_score_estimator, FDP_synth_estimator, Linked_pairs_augm, Linked_pairs: data frames (n_repeats rows x 50 thresholds, 0.50 to 0.99).

Examples

PIVs_config <- list( V1 = list(dynamics = "stable",
                               bound_mistakes = c(0.10,0.10),
                               fix_mistakes = c(NA,NA)),
                     V2 = list(dynamics = "stable",
                               bound_mistakes = c(0.10,0.10),
                               fix_mistakes = c(NA,NA)),
                     V3 = list(dynamics = "flexible",
                               bound_mistakes = c(NA,NA),
                               fix_mistakes = c(NA,NA)),
                     V4 = list(dynamics = "structured",
                               bound_mistakes = c(NA,NA),
                               fix_mistakes = c(0.03,0.03),
                               cond_hazard_cov = list(cov1=c("Xe", "Xf"),
                                                      cov2=c())) )
n_values  <- c( 5, 6, 7, 12 )
p_mistake <- list( V1 = c(0.02, 0.02), V2 = c(0.02, 0.02),
                   V3 = c(0.05, 0.05), V4 = c(0.02, 0.02) )
p_missing <- list( V1 = c(0.005, 0.005), V2 = c(0.005, 0.005),
                   V3 = c(0.005, 0.005), V4 = c(0.005, 0.005) )
cond_hazard_params <- list( V1 = c(), V2 = c(), 
                             V3 = c(), V4 = log(c(0.7, 0.6, 0.5)) )
gen_data <- simulate_data( PIVs_config, n_values, c(250, 300), 200, 
                           p_mistake, p_missing, cond_hazard_params, TRUE )
prep_data <- prepare_data( gen_data$data1, gen_data$data2, "1", "2",
                           PIVs_config, TRUE, "entity_id", TRUE )
#> '2' is the larger source, saved as B; '1' saved as A.
PIVs <- names(PIVs_config)                         
compute_augmRL_FDP_synth( "arf", prep_data$encodedA, prep_data$encodedB, PIVs, 
                          NULL, TRUE, 1, 2, "BRL", flds = PIVs, 
                          types = rep("bi",length(PIVs)) )
#> Iteration: 0, Accuracy: 48.15%
#> Warning: PIV 'V1': synthetic records out of the common support were removed.
#> Warning: PIV 'V2': synthetic records out of the common support were removed.
#> Warning: PIV 'V3': synthetic records out of the common support were removed.
#> Iteration: 0, Accuracy: 48.99%
#> Warning: PIV 'V1': synthetic records out of the common support were removed.
#> Warning: PIV 'V2': synthetic records out of the common support were removed.
#> Warning: PIV 'V4': synthetic records out of the common support were removed.
#> BRL results (average over 2 iterations) (augmented RL task) `compute_augmRL_FDP_synth`
#>                             0.50   0.55   0.60   0.65   0.70  0.75  0.80  0.85
#> FDP model score estimator   0.18   0.16   0.16   0.16   0.16  0.16  0.15  0.13
#> FDP synth data estimator    0.15   0.15   0.15   0.15   0.15  0.16  0.12  0.12
#> Linked pairs (augm. RL)   102.00  98.00  98.00  98.00  98.00 92.00 80.00 44.00
#> Linked pairs (RL)         104.00 100.00 100.00 100.00 100.00 93.00 80.00 45.00
#>                           0.90 0.95
#> FDP model score estimator 0.09    0
#> FDP synth data estimator  0.00    0
#> Linked pairs (augm. RL)   4.00    0
#> Linked pairs (RL)         4.00    0
#> $FDP_score_estimator
#>        0.50      0.51      0.52      0.53      0.54      0.55      0.56
#> 1 0.1764725 0.1701540 0.1670444 0.1639394 0.1639394 0.1639394 0.1639394
#> 2 0.1765397 0.1703452 0.1672658 0.1672658 0.1643014 0.1643014 0.1643014
#>        0.57      0.58      0.59      0.60      0.61      0.62      0.63
#> 1 0.1639394 0.1639394 0.1639394 0.1639394 0.1639394 0.1639394 0.1639394
#> 2 0.1643014 0.1643014 0.1643014 0.1643014 0.1643014 0.1643014 0.1643014
#>        0.64      0.65      0.66      0.67      0.68      0.69      0.70
#> 1 0.1639394 0.1639394 0.1639394 0.1639394 0.1623129 0.1623129 0.1623129
#> 2 0.1643014 0.1643014 0.1643014 0.1643014 0.1643014 0.1643014 0.1643014
#>        0.71      0.72     0.73      0.74      0.75      0.76      0.77
#> 1 0.1623129 0.1597338 0.157305 0.1550121 0.1527778 0.1517478 0.1517478
#> 2 0.1630000 0.1630000 0.163000 0.1599313 0.1589005 0.1541270 0.1531852
#>        0.78      0.79      0.80      0.81      0.82      0.83      0.84
#> 1 0.1500894 0.1470549 0.1418129 0.1411556 0.1366503 0.1330287 0.1297024
#> 2 0.1507407 0.1500000 0.1493595 0.1435852 0.1394444 0.1372049 0.1310897
#>        0.85      0.86      0.87      0.88       0.89       0.90       0.91 0.92
#> 1 0.1255787 0.1173203 0.1121795 0.1061438 0.09864198 0.09288889 0.08888889    0
#> 2 0.1253439 0.1202357 0.1137198 0.1062698 0.10123457 0.09527778 0.00000000    0
#>   0.93 0.94 0.95 0.96 0.97 0.98 0.99
#> 1    0    0    0    0    0    0    0
#> 2    0    0    0    0    0    0    0
#> 
#> $FDP_synth_estimator
#>         0.50      0.51      0.52      0.53      0.54      0.55      0.56
#> 1 0.09803922 0.1000000 0.1010101 0.1020408 0.1020408 0.1020408 0.1020408
#> 2 0.19417476 0.1980198 0.2000000 0.2000000 0.2020202 0.2020202 0.2020202
#>        0.57      0.58      0.59      0.60      0.61      0.62      0.63
#> 1 0.1020408 0.1020408 0.1020408 0.1020408 0.1020408 0.1020408 0.1020408
#> 2 0.2020202 0.2020202 0.2020202 0.2020202 0.2020202 0.2020202 0.2020202
#>        0.64      0.65      0.66      0.67      0.68      0.69      0.70
#> 1 0.1020408 0.1020408 0.1020408 0.1020408 0.1030928 0.1030928 0.1030928
#> 2 0.2020202 0.2020202 0.2020202 0.2020202 0.2020202 0.2020202 0.2020202
#>        0.71      0.72      0.73      0.74      0.75      0.76      0.77
#> 1 0.1030928 0.1052632 0.1075269 0.1098901 0.1123596 0.1136364 0.1136364
#> 2 0.2040816 0.2040816 0.2040816 0.2105263 0.2127660 0.2247191 0.2272727
#>        0.78      0.79      0.80      0.81      0.82      0.83      0.84
#> 1 0.1162791 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> 2 0.2352941 0.2380952 0.2409639 0.2739726 0.3030303 0.3225806 0.1960784
#>        0.85 0.86 0.87 0.88 0.89 0.90 0.91 0.92 0.93 0.94 0.95 0.96 0.97 0.98
#> 1 0.0000000    0    0    0    0    0    0    0    0    0    0    0    0    0
#> 2 0.2439024    0    0    0    0    0    0    0    0    0    0    0    0    0
#>   0.99
#> 1    0
#> 2    0
#> 
#> $Linked_pairs_augm
#>   0.50 0.51 0.52 0.53 0.54 0.55 0.56 0.57 0.58 0.59 0.60 0.61 0.62 0.63 0.64
#> 1  102  100   99   98   98   98   98   98   98   98   98   98   98   98   98
#> 2  103  101  100  100   99   99   99   99   99   99   99   99   99   99   99
#>   0.65 0.66 0.67 0.68 0.69 0.70 0.71 0.72 0.73 0.74 0.75 0.76 0.77 0.78 0.79
#> 1   98   98   98   97   97   97   97   95   93   91   89   88   88   86   83
#> 2   99   99   99   99   99   99   98   98   98   95   94   89   88   85   84
#>   0.80 0.81 0.82 0.83 0.84 0.85 0.86 0.87 0.88 0.89 0.90 0.91 0.92 0.93 0.94
#> 1   76   75   68   62   56   48   34   26   17    9    5    1    0    0    0
#> 2   83   73   66   62   51   41   33   23   14    9    4    0    0    0    0
#>   0.95 0.96 0.97 0.98 0.99
#> 1    0    0    0    0    0
#> 2    0    0    0    0    0
#> 
#> $Linked_pairs
#>   0.50 0.51 0.52 0.53 0.54 0.55 0.56 0.57 0.58 0.59 0.60 0.61 0.62 0.63 0.64
#> 1  103  101  100   99   99   99   99   99   99   99   99   99   99   99   99
#> 2  105  103  102  102  101  101  101  101  101  101  101  101  101  101  101
#>   0.65 0.66 0.67 0.68 0.69 0.70 0.71 0.72 0.73 0.74 0.75 0.76 0.77 0.78 0.79
#> 1   99   99   99   98   98   98   98   96   94   92   90   89   89   87   83
#> 2  101  101  101  101  101  101  100  100  100   97   96   91   90   87   86
#>   0.80 0.81 0.82 0.83 0.84 0.85 0.86 0.87 0.88 0.89 0.90 0.91 0.92 0.93 0.94
#> 1   76   75   68   62   56   48   34   26   17    9    5    1    0    0    0
#> 2   85   75   68   64   52   42   33   23   14    9    4    0    0    0    0
#>   0.95 0.96 0.97 0.98 0.99
#> 1    0    0    0    0    0
#> 2    0    0    0    0    0
#> 
compute_augmRL_FDP_synth( "arf", prep_data$encodedA, prep_data$encodedB, PIVs, 
                          NULL, TRUE, 1, 2, "FlexRL", data = prep_data, 
                          StEM_iter = 5, StEM_burnin = 2, 
                          gibbs_iter = 5, gibbs_burnin = 2, n_post_sample = 5 )
#> Iteration: 0, Accuracy: 48.99%
#> Warning: PIV 'V1': synthetic records out of the common support were removed.
#> Warning: PIV 'V2': synthetic records out of the common support were removed.
#> Warning: PIV 'V4': synthetic records out of the common support were removed.
#> FlexRL
#> Iteration: 0, Accuracy: 49.83%
#> Warning: PIV 'V1': synthetic records out of the common support were removed.
#> Warning: PIV 'V2': synthetic records out of the common support were removed.
#> Warning: PIV 'V3': synthetic records out of the common support were removed.
#> FlexRL
#> FlexRL results (average over 2 iterations) (augmented RL task) `compute_augmRL_FDP_synth`
#>                             0.50   0.55   0.60   0.65   0.70   0.75   0.80
#> FDP model score estimator   0.12   0.12   0.05   0.05   0.05   0.05   0.00
#> FDP synth data estimator    0.41   0.41   0.42   0.42   0.42   0.42   0.42
#> Linked pairs (augm. RL)   161.00 161.00 130.00 130.00 130.00 130.00  96.00
#> Linked pairs (RL)         168.00 168.00 136.00 136.00 136.00 136.00 100.00
#>                             0.85   0.90   0.95
#> FDP model score estimator   0.00   0.00   0.00
#> FDP synth data estimator    0.42   0.42   0.42
#> Linked pairs (augm. RL)    96.00  96.00  96.00
#> Linked pairs (RL)         100.00 100.00 100.00
#> $FDP_score_estimator
#>        0.50      0.51      0.52      0.53      0.54      0.55      0.56
#> 1 0.1067485 0.1067485 0.1067485 0.1067485 0.1067485 0.1067485 0.1067485
#> 2 0.1290698 0.1290698 0.1290698 0.1290698 0.1290698 0.1290698 0.1290698
#>        0.57      0.58      0.59       0.60       0.61       0.62       0.63
#> 1 0.1067485 0.1067485 0.1067485 0.04592593 0.04592593 0.04592593 0.04592593
#> 2 0.1290698 0.1290698 0.1290698 0.05985401 0.05985401 0.05985401 0.05985401
#>         0.64       0.65       0.66       0.67       0.68       0.69       0.70
#> 1 0.04592593 0.04592593 0.04592593 0.04592593 0.04592593 0.04592593 0.04592593
#> 2 0.05985401 0.05985401 0.05985401 0.05985401 0.05985401 0.05985401 0.05985401
#>         0.71       0.72       0.73       0.74       0.75       0.76       0.77
#> 1 0.04592593 0.04592593 0.04592593 0.04592593 0.04592593 0.04592593 0.04592593
#> 2 0.05985401 0.05985401 0.05985401 0.05985401 0.05985401 0.05985401 0.05985401
#>         0.78       0.79 0.80 0.81 0.82 0.83 0.84 0.85 0.86 0.87 0.88 0.89 0.90
#> 1 0.04592593 0.04592593    0    0    0    0    0    0    0    0    0    0    0
#> 2 0.05985401 0.05985401    0    0    0    0    0    0    0    0    0    0    0
#>   0.91 0.92 0.93 0.94 0.95 0.96 0.97 0.98 0.99
#> 1    0    0    0    0    0    0    0    0    0
#> 2    0    0    0    0    0    0    0    0    0
#> 
#> $FDP_synth_estimator
#>        0.50      0.51      0.52      0.53      0.54      0.55      0.56
#> 1 0.4487179 0.4487179 0.4487179 0.4487179 0.4487179 0.4487179 0.4487179
#> 2 0.3614458 0.3614458 0.3614458 0.3614458 0.3614458 0.3614458 0.3614458
#>        0.57      0.58      0.59      0.60      0.61      0.62      0.63
#> 1 0.4487179 0.4487179 0.4487179 0.4651163 0.4651163 0.4651163 0.4651163
#> 2 0.3614458 0.3614458 0.3614458 0.3787879 0.3787879 0.3787879 0.3787879
#>        0.64      0.65      0.66      0.67      0.68      0.69      0.70
#> 1 0.4651163 0.4651163 0.4651163 0.4651163 0.4651163 0.4651163 0.4651163
#> 2 0.3787879 0.3787879 0.3787879 0.3787879 0.3787879 0.3787879 0.3787879
#>        0.71      0.72      0.73      0.74      0.75      0.76      0.77
#> 1 0.4651163 0.4651163 0.4651163 0.4651163 0.4651163 0.4651163 0.4651163
#> 2 0.3787879 0.3787879 0.3787879 0.3787879 0.3787879 0.3787879 0.3787879
#>        0.78      0.79      0.80      0.81      0.82      0.83      0.84
#> 1 0.4651163 0.4651163 0.4000000 0.4000000 0.4000000 0.4000000 0.4000000
#> 2 0.3787879 0.3787879 0.4347826 0.4347826 0.4347826 0.4347826 0.4347826
#>        0.85      0.86      0.87      0.88      0.89      0.90      0.91
#> 1 0.4000000 0.4000000 0.4000000 0.4000000 0.4000000 0.4000000 0.4000000
#> 2 0.4347826 0.4347826 0.4347826 0.4347826 0.4347826 0.4347826 0.4347826
#>        0.92      0.93      0.94      0.95      0.96      0.97      0.98
#> 1 0.4000000 0.4000000 0.4000000 0.4000000 0.4000000 0.4000000 0.4000000
#> 2 0.4347826 0.4347826 0.4347826 0.4347826 0.4347826 0.4347826 0.4347826
#>        0.99
#> 1 0.4000000
#> 2 0.4347826
#> 
#> $Linked_pairs_augm
#>   0.50 0.51 0.52 0.53 0.54 0.55 0.56 0.57 0.58 0.59 0.60 0.61 0.62 0.63 0.64
#> 1  156  156  156  156  156  156  156  156  156  156  129  129  129  129  129
#> 2  166  166  166  166  166  166  166  166  166  166  132  132  132  132  132
#>   0.65 0.66 0.67 0.68 0.69 0.70 0.71 0.72 0.73 0.74 0.75 0.76 0.77 0.78 0.79
#> 1  129  129  129  129  129  129  129  129  129  129  129  129  129  129  129
#> 2  132  132  132  132  132  132  132  132  132  132  132  132  132  132  132
#>   0.80 0.81 0.82 0.83 0.84 0.85 0.86 0.87 0.88 0.89 0.90 0.91 0.92 0.93 0.94
#> 1  100  100  100  100  100  100  100  100  100  100  100  100  100  100  100
#> 2   92   92   92   92   92   92   92   92   92   92   92   92   92   92   92
#>   0.95 0.96 0.97 0.98 0.99
#> 1  100  100  100  100  100
#> 2   92   92   92   92   92
#> 
#> $Linked_pairs
#>   0.50 0.51 0.52 0.53 0.54 0.55 0.56 0.57 0.58 0.59 0.60 0.61 0.62 0.63 0.64
#> 1  163  163  163  163  163  163  163  163  163  163  135  135  135  135  135
#> 2  172  172  172  172  172  172  172  172  172  172  137  137  137  137  137
#>   0.65 0.66 0.67 0.68 0.69 0.70 0.71 0.72 0.73 0.74 0.75 0.76 0.77 0.78 0.79
#> 1  135  135  135  135  135  135  135  135  135  135  135  135  135  135  135
#> 2  137  137  137  137  137  137  137  137  137  137  137  137  137  137  137
#>   0.80 0.81 0.82 0.83 0.84 0.85 0.86 0.87 0.88 0.89 0.90 0.91 0.92 0.93 0.94
#> 1  104  104  104  104  104  104  104  104  104  104  104  104  104  104  104
#> 2   96   96   96   96   96   96   96   96   96   96   96   96   96   96   96
#>   0.95 0.96 0.97 0.98 0.99
#> 1  104  104  104  104  104
#> 2   96   96   96   96   96
#>