Estimate the false discovery proportion of a record-linkage method via synthetic augmentation

Repeatedly augments file B with synthetic records (synthesise()), runs the chosen record-linkage method link_with_*, and compares the synthetic-vs-real proportion among linked pairs to estimate the false discovery proportion, for a range of score thresholds (0.50 to 0.99). See https://doi.org/10.1002/sim.70292 for the method.

compute_augmRL_FDP_synth(
  synth_method,
  encodedA,
  encodedB,
  PIVs,
  n_synth = NULL,
  restrict_support_intersection = TRUE,
  maxIter4CV = 10,
  n_repeats = 10,
  RL_method,
  ...
)

Arguments

synth_method

Passed to synthesise(): "arf", "synthpop", or "mice".

encodedA, encodedB

The two encoded data sources as returned by prepare_data() (encodedA is the smaller one).

PIVs

Character vector, names of the PIVs.

n_synth

Integer, number of synthetic records to generate per iteration (default: 10% of nrow(encodedB)).

restrict_support_intersection

Passed to synthesise().

maxIter4CV

Integer, max number of retries per iteration if no valid FDP estimate is obtained.

n_repeats

Integer, number of augmentation iterations to average over.

RL_method

One of "multilink", "fastLink", "BRL", "reclin2", "diyar", "fedmatch", "FlexRL".

...

Extra arguments forwarded to the chosen link_with_* wrapper (i.e. to the underlying record-linkage package).

Value

List with FDP_score_estimator, FDP_synth_estimator, Linked_pairs_augm, Linked_pairs: data frames (n_repeats rows x 50 thresholds, 0.50 to 0.99).

Examples

PIVs_config <- list( V1 = list(dynamics = "stable",
                               bound_mistakes = c(0.10,0.10),
                               fix_mistakes = c(NA,NA)),
                     V2 = list(dynamics = "stable",
                               bound_mistakes = c(0.10,0.10),
                               fix_mistakes = c(NA,NA)),
                     V3 = list(dynamics = "flexible",
                               bound_mistakes = c(NA,NA),
                               fix_mistakes = c(NA,NA)),
                     V4 = list(dynamics = "structured",
                               bound_mistakes = c(NA,NA),
                               fix_mistakes = c(0.03,0.03),
                               cond_hazard_cov = list(cov1=c("Xe", "Xf"),
                                                      cov2=c())) )
n_values  <- c( 5, 6, 7, 12 )
p_mistake <- list( V1 = c(0.02, 0.02), V2 = c(0.02, 0.02),
                   V3 = c(0.05, 0.05), V4 = c(0.02, 0.02) )
p_missing <- list( V1 = c(0.005, 0.005), V2 = c(0.005, 0.005),
                   V3 = c(0.005, 0.005), V4 = c(0.005, 0.005) )
cond_hazard_params <- list( V1 = c(), V2 = c(), 
                             V3 = c(), V4 = log(c(0.7, 0.6, 0.5)) )
gen_data <- simulate_data( PIVs_config, n_values, c(250, 300), 200, 
                           p_mistake, p_missing, cond_hazard_params, TRUE )
prep_data <- prepare_data( gen_data$data1, gen_data$data2, "1", "2",
                           PIVs_config, TRUE, "entity_id", TRUE )
#> '2' is the larger source, saved as B; '1' saved as A.
PIVs <- names(PIVs_config)                         
compute_augmRL_FDP_synth( "arf", prep_data$encodedA, prep_data$encodedB, PIVs, 
                          NULL, TRUE, 1, 2, "BRL", flds = PIVs, 
                          types = rep("bi",length(PIVs)) )
#> Iteration: 0, Accuracy: 50.51%
#> Iteration: 1, Accuracy: 39.39%
#> Warning: PIV 'V3': synthetic records out of the common support were removed.
#> Iteration: 0, Accuracy: 49.32%
#> Warning: PIV 'V3': synthetic records out of the common support were removed.
#> BRL results (average over 2 iterations) (augmented RL task) `compute_augmRL_FDP_synth`
#> Decision rule threshold   	0.5	0.55	0.6	0.65	0.7	0.75	0.8	0.85	0.9	0.95
#> FDP model score estimator 	0.18	0.17	0.17	0.17	0.17	0.17	0.16	0.14	0	0
#> FDP synth data estimator  	0.24	0.24	0.24	0.24	0.24	0.18	0.07	0	0	0
#> Linked pairs (augm. RL)   	85	82	82	82	82	82	69	24	0	0
#> Linked pairs (RL)         	87	84	84	84	84	84	70	24	0	0
#> $FDP_score_estimator
#>        0.50      0.51      0.52      0.53      0.54      0.55      0.56
#> 1 0.1779716 0.1741961 0.1741961 0.1741961 0.1741961 0.1708201 0.1708201
#> 2 0.1791540 0.1755683 0.1682484 0.1682484 0.1682484 0.1682484 0.1682484
#>        0.57      0.58      0.59      0.60      0.61      0.62      0.63
#> 1 0.1708201 0.1708201 0.1708201 0.1708201 0.1708201 0.1708201 0.1708201
#> 2 0.1682484 0.1682484 0.1682484 0.1682484 0.1682484 0.1682484 0.1682484
#>        0.64      0.65      0.66      0.67      0.68      0.69      0.70
#> 1 0.1708201 0.1708201 0.1708201 0.1708201 0.1708201 0.1708201 0.1708201
#> 2 0.1682484 0.1682484 0.1682484 0.1682484 0.1682484 0.1682484 0.1682484
#>        0.71      0.72      0.73      0.74      0.75      0.76      0.77
#> 1 0.1708201 0.1708201 0.1708201 0.1708201 0.1698126 0.1698126 0.1682167
#> 2 0.1682484 0.1682484 0.1682484 0.1682484 0.1672354 0.1652981 0.1652981
#>        0.78      0.79      0.80      0.81      0.82      0.83      0.84
#> 1 0.1644444 0.1624810 0.1601771 0.1563620 0.1529091 0.1489855 0.1433007
#> 2 0.1637917 0.1605185 0.1576190 0.1559867 0.1527222 0.1468322 0.1418919
#>        0.85      0.86      0.87      0.88      0.89 0.90 0.91 0.92 0.93 0.94
#> 1 0.1374242 0.1315152 0.1207407 0.1100000 0.0000000    0    0    0    0    0
#> 2 0.1355556 0.1314583 0.1228889 0.1055556 0.1055556    0    0    0    0    0
#>   0.95 0.96 0.97 0.98 0.99
#> 1    0    0    0    0    0
#> 2    0    0    0    0    0
#> 
#> $FDP_synth_estimator
#>        0.50      0.51      0.52      0.53      0.54      0.55      0.56
#> 1 0.2380952 0.2409639 0.2409639 0.2409639 0.2409639 0.2439024 0.2439024
#> 2 0.2325581 0.2352941 0.2409639 0.2409639 0.2409639 0.2409639 0.2409639
#>        0.57      0.58      0.59      0.60      0.61      0.62      0.63
#> 1 0.2439024 0.2439024 0.2439024 0.2439024 0.2439024 0.2439024 0.2439024
#> 2 0.2409639 0.2409639 0.2409639 0.2409639 0.2409639 0.2409639 0.2409639
#>        0.64      0.65      0.66      0.67      0.68      0.69      0.70
#> 1 0.2439024 0.2439024 0.2439024 0.2439024 0.2439024 0.2439024 0.2439024
#> 2 0.2409639 0.2409639 0.2409639 0.2409639 0.2409639 0.2409639 0.2409639
#>        0.71      0.72      0.73      0.74      0.75      0.76      0.77
#> 1 0.2439024 0.2439024 0.2439024 0.2439024 0.2469136 0.2469136 0.1250000
#> 2 0.2409639 0.2409639 0.2409639 0.2409639 0.1204819 0.1234568 0.1234568
#>        0.78      0.79      0.80      0.81      0.82      0.83 0.84 0.85 0.86
#> 1 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000    0    0    0
#> 2 0.1265823 0.1351351 0.1449275 0.1515152 0.1694915 0.2173913    0    0    0
#>   0.87 0.88 0.89 0.90 0.91 0.92 0.93 0.94 0.95 0.96 0.97 0.98 0.99
#> 1    0    0    0    0    0    0    0    0    0    0    0    0    0
#> 2    0    0    0    0    0    0    0    0    0    0    0    0    0
#> 
#> $Linked_pairs_augm
#>   0.50 0.51 0.52 0.53 0.54 0.55 0.56 0.57 0.58 0.59 0.60 0.61 0.62 0.63 0.64
#> 1   84   83   83   83   83   82   82   82   82   82   82   82   82   82   82
#> 2   86   85   83   83   83   83   83   83   83   83   83   83   83   83   83
#>   0.65 0.66 0.67 0.68 0.69 0.70 0.71 0.72 0.73 0.74 0.75 0.76 0.77 0.78 0.79
#> 1   82   82   82   82   82   82   82   82   82   82   81   81   80   76   73
#> 2   83   83   83   83   83   83   83   83   83   83   83   81   81   79   74
#>   0.80 0.81 0.82 0.83 0.84 0.85 0.86 0.87 0.88 0.89 0.90 0.91 0.92 0.93 0.94
#> 1   69   62   55   46   34   22   11    3    1    0    0    0    0    0    0
#> 2   69   66   59   46   37   25   16    5    1    1    0    0    0    0    0
#>   0.95 0.96 0.97 0.98 0.99
#> 1    0    0    0    0    0
#> 2    0    0    0    0    0
#> 
#> $Linked_pairs
#>   0.50 0.51 0.52 0.53 0.54 0.55 0.56 0.57 0.58 0.59 0.60 0.61 0.62 0.63 0.64
#> 1   86   85   85   85   85   84   84   84   84   84   84   84   84   84   84
#> 2   88   87   85   85   85   85   85   85   85   85   85   85   85   85   85
#>   0.65 0.66 0.67 0.68 0.69 0.70 0.71 0.72 0.73 0.74 0.75 0.76 0.77 0.78 0.79
#> 1   84   84   84   84   84   84   84   84   84   84   83   83   81   76   73
#> 2   85   85   85   85   85   85   85   85   85   85   84   82   82   80   75
#>   0.80 0.81 0.82 0.83 0.84 0.85 0.86 0.87 0.88 0.89 0.90 0.91 0.92 0.93 0.94
#> 1   69   62   55   46   34   22   11    3    1    0    0    0    0    0    0
#> 2   70   67   60   47   37   25   16    5    1    1    0    0    0    0    0
#>   0.95 0.96 0.97 0.98 0.99
#> 1    0    0    0    0    0
#> 2    0    0    0    0    0
#> 
compute_augmRL_FDP_synth( "arf", prep_data$encodedA, prep_data$encodedB, PIVs, 
                          NULL, TRUE, 1, 2, "FlexRL", data = prep_data, 
                          StEM_iter = 5, StEM_burnin = 2, 
                          gibbs_iter = 5, gibbs_burnin = 2, n_post_sample = 5 )
#> Iteration: 0, Accuracy: 49.32%
#> Warning: PIV 'V3': synthetic records out of the common support were removed.
#> FlexRL
#> Iteration: 0, Accuracy: 46.37%
#> Warning: PIV 'V2': synthetic records out of the common support were removed.
#> Warning: PIV 'V3': synthetic records out of the common support were removed.
#> FlexRL
#> FlexRL results (average over 2 iterations) (augmented RL task) `compute_augmRL_FDP_synth`
#> Decision rule threshold   	0.5	0.55	0.6	0.65	0.7	0.75	0.8	0.85	0.9	0.95
#> FDP model score estimator 	0.15	0.15	0.05	0.05	0.05	0.05	0	0	0	0
#> FDP synth data estimator  	0.35	0.35	0.36	0.36	0.36	0.36	0.21	0.21	0.21	0.21
#> Linked pairs (augm. RL)   	141	141	100	100	100	100	74	74	74	74
#> Linked pairs (RL)         	146	146	104	104	104	104	76	76	76	76
#> $FDP_score_estimator
#>        0.50      0.51      0.52      0.53      0.54      0.55      0.56
#> 1 0.1555556 0.1555556 0.1555556 0.1555556 0.1555556 0.1555556 0.1555556
#> 2 0.1513514 0.1513514 0.1513514 0.1513514 0.1513514 0.1513514 0.1513514
#>        0.57      0.58      0.59       0.60       0.61       0.62       0.63
#> 1 0.1555556 0.1555556 0.1555556 0.04800000 0.04800000 0.04800000 0.04800000
#> 2 0.1513514 0.1513514 0.1513514 0.05925926 0.05925926 0.05925926 0.05925926
#>         0.64       0.65       0.66       0.67       0.68       0.69       0.70
#> 1 0.04800000 0.04800000 0.04800000 0.04800000 0.04800000 0.04800000 0.04800000
#> 2 0.05925926 0.05925926 0.05925926 0.05925926 0.05925926 0.05925926 0.05925926
#>         0.71       0.72       0.73       0.74       0.75       0.76       0.77
#> 1 0.04800000 0.04800000 0.04800000 0.04800000 0.04800000 0.04800000 0.04800000
#> 2 0.05925926 0.05925926 0.05925926 0.05925926 0.05925926 0.05925926 0.05925926
#>         0.78       0.79 0.80 0.81 0.82 0.83 0.84 0.85 0.86 0.87 0.88 0.89 0.90
#> 1 0.04800000 0.04800000    0    0    0    0    0    0    0    0    0    0    0
#> 2 0.05925926 0.05925926    0    0    0    0    0    0    0    0    0    0    0
#>   0.91 0.92 0.93 0.94 0.95 0.96 0.97 0.98 0.99
#> 1    0    0    0    0    0    0    0    0    0
#> 2    0    0    0    0    0    0    0    0    0
#> 
#> $FDP_synth_estimator
#>        0.50      0.51      0.52      0.53      0.54      0.55      0.56
#> 1 0.3597122 0.3597122 0.3597122 0.3597122 0.3597122 0.3597122 0.3597122
#> 2 0.3496503 0.3496503 0.3496503 0.3496503 0.3496503 0.3496503 0.3496503
#>        0.57      0.58      0.59      0.60      0.61      0.62      0.63
#> 1 0.3597122 0.3597122 0.3597122 0.5263158 0.5263158 0.5263158 0.5263158
#> 2 0.3496503 0.3496503 0.3496503 0.1886792 0.1886792 0.1886792 0.1886792
#>        0.64      0.65      0.66      0.67      0.68      0.69      0.70
#> 1 0.5263158 0.5263158 0.5263158 0.5263158 0.5263158 0.5263158 0.5263158
#> 2 0.1886792 0.1886792 0.1886792 0.1886792 0.1886792 0.1886792 0.1886792
#>        0.71      0.72      0.73      0.74      0.75      0.76      0.77
#> 1 0.5263158 0.5263158 0.5263158 0.5263158 0.5263158 0.5263158 0.5263158
#> 2 0.1886792 0.1886792 0.1886792 0.1886792 0.1886792 0.1886792 0.1886792
#>        0.78      0.79      0.80      0.81      0.82      0.83      0.84
#> 1 0.5263158 0.5263158 0.4109589 0.4109589 0.4109589 0.4109589 0.4109589
#> 2 0.1886792 0.1886792 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>        0.85      0.86      0.87      0.88      0.89      0.90      0.91
#> 1 0.4109589 0.4109589 0.4109589 0.4109589 0.4109589 0.4109589 0.4109589
#> 2 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>        0.92      0.93      0.94      0.95      0.96      0.97      0.98
#> 1 0.4109589 0.4109589 0.4109589 0.4109589 0.4109589 0.4109589 0.4109589
#> 2 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>        0.99
#> 1 0.4109589
#> 2 0.0000000
#> 
#> $Linked_pairs_augm
#>   0.50 0.51 0.52 0.53 0.54 0.55 0.56 0.57 0.58 0.59 0.60 0.61 0.62 0.63 0.64
#> 1  139  139  139  139  139  139  139  139  139  139   95   95   95   95   95
#> 2  143  143  143  143  143  143  143  143  143  143  106  106  106  106  106
#>   0.65 0.66 0.67 0.68 0.69 0.70 0.71 0.72 0.73 0.74 0.75 0.76 0.77 0.78 0.79
#> 1   95   95   95   95   95   95   95   95   95   95   95   95   95   95   95
#> 2  106  106  106  106  106  106  106  106  106  106  106  106  106  106  106
#>   0.80 0.81 0.82 0.83 0.84 0.85 0.86 0.87 0.88 0.89 0.90 0.91 0.92 0.93 0.94
#> 1   73   73   73   73   73   73   73   73   73   73   73   73   73   73   73
#> 2   76   76   76   76   76   76   76   76   76   76   76   76   76   76   76
#>   0.95 0.96 0.97 0.98 0.99
#> 1   73   73   73   73   73
#> 2   76   76   76   76   76
#> 
#> $Linked_pairs
#>   0.50 0.51 0.52 0.53 0.54 0.55 0.56 0.57 0.58 0.59 0.60 0.61 0.62 0.63 0.64
#> 1  144  144  144  144  144  144  144  144  144  144  100  100  100  100  100
#> 2  148  148  148  148  148  148  148  148  148  148  108  108  108  108  108
#>   0.65 0.66 0.67 0.68 0.69 0.70 0.71 0.72 0.73 0.74 0.75 0.76 0.77 0.78 0.79
#> 1  100  100  100  100  100  100  100  100  100  100  100  100  100  100  100
#> 2  108  108  108  108  108  108  108  108  108  108  108  108  108  108  108
#>   0.80 0.81 0.82 0.83 0.84 0.85 0.86 0.87 0.88 0.89 0.90 0.91 0.92 0.93 0.94
#> 1   76   76   76   76   76   76   76   76   76   76   76   76   76   76   76
#> 2   76   76   76   76   76   76   76   76   76   76   76   76   76   76   76
#>   0.95 0.96 0.97 0.98 0.99
#> 1   76   76   76   76   76
#> 2   76   76   76   76   76
#>