Estimate the false discovery proportion of a record-linkage method via synthetic augmentation

Repeatedly augments file B with synthetic records (synthesise()), runs the chosen record-linkage method link_with_*, and compares the synthetic-vs-real proportion among linked pairs to estimate the false discovery proportion, for a range of score thresholds (0.50 to 0.99). See https://doi.org/10.1002/sim.70292 for the method.

compute_augmRL_FDP_synth(
  synth_method,
  encodedA,
  encodedB,
  PIVs,
  n_synth = NULL,
  restrict_support_intersection = TRUE,
  maxIter4CV = 10,
  n_repeats = 10,
  RL_method,
  ...
)

Arguments

synth_method

Passed to synthesise(): "arf", "synthpop", or "mice".

encodedA, encodedB

The two encoded data sources as returned by prepare_data() (encodedA is the smaller one).

PIVs

Character vector, names of the PIVs.

n_synth

Integer, number of synthetic records to generate per iteration (default: 10% of nrow(encodedB)).

restrict_support_intersection

Passed to synthesise().

maxIter4CV

Integer, max number of retries per iteration if no valid FDP estimate is obtained.

n_repeats

Integer, number of augmentation iterations to average over.

RL_method

One of "multilink", "fastLink", "BRL", "reclin2", "diyar", "fedmatch", "FlexRL".

...

Extra arguments forwarded to the chosen link_with_* wrapper (i.e. to the underlying record-linkage package).

Value

List with FDP_score_estimator, FDP_synth_estimator, Linked_pairs_augm, Linked_pairs: data frames (n_repeats rows x 50 thresholds, 0.50 to 0.99).

Examples

PIVs_config <- list( V1 = list(dynamics = "stable",
                               bound_mistakes = c(0.10,0.10),
                               fix_mistakes = c(NA,NA)),
                     V2 = list(dynamics = "stable",
                               bound_mistakes = c(0.10,0.10),
                               fix_mistakes = c(NA,NA)),
                     V3 = list(dynamics = "flexible",
                               bound_mistakes = c(NA,NA),
                               fix_mistakes = c(NA,NA)),
                     V4 = list(dynamics = "structured",
                               bound_mistakes = c(NA,NA),
                               fix_mistakes = c(0.03,0.03),
                               cond_hazard_cov = list(cov1=c("Xe", "Xf"),
                                                      cov2=c())) )
n_values  <- c( 5, 6, 7, 12 )
p_mistake <- list( V1 = c(0.02, 0.02), V2 = c(0.02, 0.02),
                   V3 = c(0.05, 0.05), V4 = c(0.02, 0.02) )
p_missing <- list( V1 = c(0.005, 0.005), V2 = c(0.005, 0.005),
                   V3 = c(0.005, 0.005), V4 = c(0.005, 0.005) )
cond_hazard_params <- list( V1 = c(), V2 = c(), 
                             V3 = c(), V4 = log(c(0.7, 0.6, 0.5)) )
gen_data <- simulate_data( PIVs_config, n_values, c(150, 200), 100, 
                           p_mistake, p_missing, cond_hazard_params, TRUE )
prep_data <- prepare_data( gen_data$data1, gen_data$data2, "1", "2",
                           PIVs_config, TRUE, "entity_id", TRUE )
#> Warning: PIV 'V4': values out of the common support in data2 [5].
#> '2' is the larger source, saved as B; '1' saved as A.
PIVs <- names(PIVs_config)                         
compute_augmRL_FDP_synth( "arf", prep_data$encodedA, prep_data$encodedB, PIVs, 
                          NULL, TRUE, 1, 2, "BRL", flds = PIVs, 
                          types = rep("bi",length(PIVs)) )
#> Iteration: 0, Accuracy: 50%
#> Warning: PIV 'V3': synthetic records out of the common support were removed.
#> Warning: PIV 'V4': synthetic records out of the common support were removed.
#> Iteration: 0, Accuracy: 49.1%
#> Warning: PIV 'V1': synthetic records out of the common support were removed.
#> Warning: PIV 'V3': synthetic records out of the common support were removed.
#> Warning: PIV 'V4': synthetic records out of the common support were removed.
#> BRL results (average over 2 iterations) (augmented RL task) `compute_augmRL_FDP_synth`
#> Decision rule threshold   	0.5	0.55	0.6	0.65	0.7	0.75	0.8	0.85	0.9	0.95
#> FDP model score estimator 	0.21	0.21	0.21	0.2	0.2	0.19	0.17	0.14	0	0
#> FDP synth data estimator  	0	0	0	0	0	0	0	0	0	0
#> Linked pairs (augm. RL)   	46	46	46	44	42	36	21	4	0	0
#> Linked pairs (RL)         	46	46	46	44	42	36	21	4	0	0
#> $FDP_score_estimator
#>        0.50      0.51      0.52      0.53      0.54      0.55      0.56
#> 1 0.2108454 0.2108454 0.2108454 0.2108454 0.2108454 0.2108454 0.2108454
#> 2 0.2122222 0.2122222 0.2122222 0.2122222 0.2122222 0.2122222 0.2122222
#>        0.57      0.58      0.59      0.60      0.61      0.62      0.63
#> 1 0.2108454 0.2108454 0.2108454 0.2108454 0.2108454 0.2108454 0.2108454
#> 2 0.2122222 0.2122222 0.2122222 0.2122222 0.2122222 0.2122222 0.2122222
#>        0.64      0.65      0.66      0.67      0.68      0.69      0.70
#> 1 0.2075062 0.2040657 0.2040657 0.2010078 0.2010078 0.1984127 0.1984127
#> 2 0.2122222 0.2058838 0.2058838 0.2058838 0.2031266 0.2031266 0.2006614
#>        0.71      0.72      0.73      0.74      0.75      0.76      0.77
#> 1 0.1984127 0.1984127 0.1946111 0.1910526 0.1857143 0.1823906 0.1807986
#> 2 0.2006614 0.1987263 0.1987263 0.1970278 0.1924324 0.1894921 0.1880392
#>        0.78      0.79      0.80      0.81      0.82      0.83      0.84
#> 1 0.1747222 0.1698222 0.1645960 0.1602339 0.1572549 0.1487879 0.1426389
#> 2 0.1832593 0.1785897 0.1706667 0.1652778 0.1586111 0.1558889 0.1520635
#>        0.85      0.86      0.87 0.88 0.89 0.90 0.91 0.92 0.93 0.94 0.95 0.96
#> 1 0.1388889 0.1294444 0.1211111    0    0    0    0    0    0    0    0    0
#> 2 0.1405556 0.1322222 0.0000000    0    0    0    0    0    0    0    0    0
#>   0.97 0.98 0.99
#> 1    0    0    0
#> 2    0    0    0
#> 
#> $FDP_synth_estimator
#>   0.50 0.51 0.52 0.53 0.54 0.55 0.56 0.57 0.58 0.59 0.60 0.61 0.62 0.63 0.64
#> 1    0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> 2    0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#>   0.65 0.66 0.67 0.68 0.69 0.70 0.71 0.72 0.73 0.74 0.75 0.76 0.77 0.78 0.79
#> 1    0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> 2    0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#>   0.80 0.81 0.82 0.83 0.84 0.85 0.86 0.87 0.88 0.89 0.90 0.91 0.92 0.93 0.94
#> 1    0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#> 2    0    0    0    0    0    0    0    0    0    0    0    0    0    0    0
#>   0.95 0.96 0.97 0.98 0.99
#> 1    0    0    0    0    0
#> 2    0    0    0    0    0
#> 
#> $Linked_pairs_augm
#>   0.50 0.51 0.52 0.53 0.54 0.55 0.56 0.57 0.58 0.59 0.60 0.61 0.62 0.63 0.64
#> 1   46   46   46   46   46   46   46   46   46   46   46   46   46   46   45
#> 2   46   46   46   46   46   46   46   46   46   46   46   46   46   46   46
#>   0.65 0.66 0.67 0.68 0.69 0.70 0.71 0.72 0.73 0.74 0.75 0.76 0.77 0.78 0.79
#> 1   44   44   43   43   42   42   42   42   40   38   35   33   32   28   25
#> 2   44   44   44   43   43   42   42   41   41   40   37   35   34   30   26
#>   0.80 0.81 0.82 0.83 0.84 0.85 0.86 0.87 0.88 0.89 0.90 0.91 0.92 0.93 0.94
#> 1   22   19   17   11    8    6    2    1    0    0    0    0    0    0    0
#> 2   20   16   12   10    7    2    1    0    0    0    0    0    0    0    0
#>   0.95 0.96 0.97 0.98 0.99
#> 1    0    0    0    0    0
#> 2    0    0    0    0    0
#> 
#> $Linked_pairs
#>   0.50 0.51 0.52 0.53 0.54 0.55 0.56 0.57 0.58 0.59 0.60 0.61 0.62 0.63 0.64
#> 1   46   46   46   46   46   46   46   46   46   46   46   46   46   46   45
#> 2   46   46   46   46   46   46   46   46   46   46   46   46   46   46   46
#>   0.65 0.66 0.67 0.68 0.69 0.70 0.71 0.72 0.73 0.74 0.75 0.76 0.77 0.78 0.79
#> 1   44   44   43   43   42   42   42   42   40   38   35   33   32   28   25
#> 2   44   44   44   43   43   42   42   41   41   40   37   35   34   30   26
#>   0.80 0.81 0.82 0.83 0.84 0.85 0.86 0.87 0.88 0.89 0.90 0.91 0.92 0.93 0.94
#> 1   22   19   17   11    8    6    2    1    0    0    0    0    0    0    0
#> 2   20   16   12   10    7    2    1    0    0    0    0    0    0    0    0
#>   0.95 0.96 0.97 0.98 0.99
#> 1    0    0    0    0    0
#> 2    0    0    0    0    0
#> 
compute_augmRL_FDP_synth( "arf", prep_data$encodedA, prep_data$encodedB, PIVs, 
                          NULL, TRUE, 1, 2, "FlexRL", data = prep_data, 
                          StEM_iter = 5, StEM_burnin = 2, 
                          gibbs_iter = 5, gibbs_burnin = 2, n_post_sample = 5 )
#> Iteration: 0, Accuracy: 49.1%
#> Warning: PIV 'V1': synthetic records out of the common support were removed.
#> Warning: PIV 'V3': synthetic records out of the common support were removed.
#> Warning: PIV 'V4': synthetic records out of the common support were removed.
#> FlexRL
#> Iteration: 0, Accuracy: 47.07%
#> Warning: PIV 'V2': synthetic records out of the common support were removed.
#> Warning: PIV 'V3': synthetic records out of the common support were removed.
#> Warning: PIV 'V4': synthetic records out of the common support were removed.
#> FlexRL
#> FlexRL results (average over 2 iterations) (augmented RL task) `compute_augmRL_FDP_synth`
#> Decision rule threshold   	0.5	0.55	0.6	0.65	0.7	0.75	0.8	0.85	0.9	0.95
#> FDP model score estimator 	0.15	0.15	0.06	0.06	0.06	0.06	0	0	0	0
#> FDP synth data estimator  	0.18	0.18	0.09	0.09	0.09	0.09	0.12	0.12	0.12	0.12
#> Linked pairs (augm. RL)   	90	90	65	65	65	65	46	46	46	46
#> Linked pairs (RL)         	91	91	66	66	66	66	46	46	46	46
#> $FDP_score_estimator
#>        0.50      0.51      0.52      0.53      0.54      0.55      0.56
#> 1 0.1400000 0.1400000 0.1400000 0.1400000 0.1400000 0.1400000 0.1400000
#> 2 0.1695652 0.1695652 0.1695652 0.1695652 0.1695652 0.1695652 0.1695652
#>        0.57      0.58      0.59       0.60       0.61       0.62       0.63
#> 1 0.1400000 0.1400000 0.1400000 0.06086957 0.06086957 0.06086957 0.06086957
#> 2 0.1695652 0.1695652 0.1695652 0.05806452 0.05806452 0.05806452 0.05806452
#>         0.64       0.65       0.66       0.67       0.68       0.69       0.70
#> 1 0.06086957 0.06086957 0.06086957 0.06086957 0.06086957 0.06086957 0.06086957
#> 2 0.05806452 0.05806452 0.05806452 0.05806452 0.05806452 0.05806452 0.05806452
#>         0.71       0.72       0.73       0.74       0.75       0.76       0.77
#> 1 0.06086957 0.06086957 0.06086957 0.06086957 0.06086957 0.06086957 0.06086957
#> 2 0.05806452 0.05806452 0.05806452 0.05806452 0.05806452 0.05806452 0.05806452
#>         0.78       0.79 0.80 0.81 0.82 0.83 0.84 0.85 0.86 0.87 0.88 0.89 0.90
#> 1 0.06086957 0.06086957    0    0    0    0    0    0    0    0    0    0    0
#> 2 0.05806452 0.05806452    0    0    0    0    0    0    0    0    0    0    0
#>   0.91 0.92 0.93 0.94 0.95 0.96 0.97 0.98 0.99
#> 1    0    0    0    0    0    0    0    0    0
#> 2    0    0    0    0    0    0    0    0    0
#> 
#> $FDP_synth_estimator
#>        0.50      0.51      0.52      0.53      0.54      0.55      0.56
#> 1 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> 2 0.3512714 0.3512714 0.3512714 0.3512714 0.3512714 0.3512714 0.3512714
#>        0.57      0.58      0.59      0.60      0.61      0.62      0.63
#> 1 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> 2 0.3512714 0.3512714 0.3512714 0.1708369 0.1708369 0.1708369 0.1708369
#>        0.64      0.65      0.66      0.67      0.68      0.69      0.70
#> 1 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> 2 0.1708369 0.1708369 0.1708369 0.1708369 0.1708369 0.1708369 0.1708369
#>        0.71      0.72      0.73      0.74      0.75      0.76      0.77
#> 1 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> 2 0.1708369 0.1708369 0.1708369 0.1708369 0.1708369 0.1708369 0.1708369
#>        0.78      0.79      0.80      0.81      0.82      0.83      0.84
#> 1 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> 2 0.1708369 0.1708369 0.2423501 0.2423501 0.2423501 0.2423501 0.2423501
#>        0.85      0.86      0.87      0.88      0.89      0.90      0.91
#> 1 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> 2 0.2423501 0.2423501 0.2423501 0.2423501 0.2423501 0.2423501 0.2423501
#>        0.92      0.93      0.94      0.95      0.96      0.97      0.98
#> 1 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> 2 0.2423501 0.2423501 0.2423501 0.2423501 0.2423501 0.2423501 0.2423501
#>        0.99
#> 1 0.0000000
#> 2 0.2423501
#> 
#> $Linked_pairs_augm
#>   0.50 0.51 0.52 0.53 0.54 0.55 0.56 0.57 0.58 0.59 0.60 0.61 0.62 0.63 0.64
#> 1   90   90   90   90   90   90   90   90   90   90   69   69   69   69   69
#> 2   89   89   89   89   89   89   89   89   89   89   61   61   61   61   61
#>   0.65 0.66 0.67 0.68 0.69 0.70 0.71 0.72 0.73 0.74 0.75 0.76 0.77 0.78 0.79
#> 1   69   69   69   69   69   69   69   69   69   69   69   69   69   69   69
#> 2   61   61   61   61   61   61   61   61   61   61   61   61   61   61   61
#>   0.80 0.81 0.82 0.83 0.84 0.85 0.86 0.87 0.88 0.89 0.90 0.91 0.92 0.93 0.94
#> 1   48   48   48   48   48   48   48   48   48   48   48   48   48   48   48
#> 2   43   43   43   43   43   43   43   43   43   43   43   43   43   43   43
#>   0.95 0.96 0.97 0.98 0.99
#> 1   48   48   48   48   48
#> 2   43   43   43   43   43
#> 
#> $Linked_pairs
#>   0.50 0.51 0.52 0.53 0.54 0.55 0.56 0.57 0.58 0.59 0.60 0.61 0.62 0.63 0.64
#> 1   90   90   90   90   90   90   90   90   90   90   69   69   69   69   69
#> 2   92   92   92   92   92   92   92   92   92   92   62   62   62   62   62
#>   0.65 0.66 0.67 0.68 0.69 0.70 0.71 0.72 0.73 0.74 0.75 0.76 0.77 0.78 0.79
#> 1   69   69   69   69   69   69   69   69   69   69   69   69   69   69   69
#> 2   62   62   62   62   62   62   62   62   62   62   62   62   62   62   62
#>   0.80 0.81 0.82 0.83 0.84 0.85 0.86 0.87 0.88 0.89 0.90 0.91 0.92 0.93 0.94
#> 1   48   48   48   48   48   48   48   48   48   48   48   48   48   48   48
#> 2   44   44   44   44   44   44   44   44   44   44   44   44   44   44   44
#>   0.95 0.96 0.97 0.98 0.99
#> 1   48   48   48   48   48
#> 2   44   44   44   44   44
#>